+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -2,6 +2,8 @@ Version 5.2.1
 =============
 
  * 2015-10-16 Context.getParameterFileItems() now returns Map<String, List<FileItem>> (jjlauer)
+ * 2015-10-18 OptionalBinder for ObjectMapper and XmlMapper, so that users can easily override 
+              both of them (amit2103)
 
 Version 5.2.0
 =============